mac64 gcc4.2编译通过
1、处理各种警告及测试用例失败问题［64位］
2、ConvertUTF，UTF32固定为32位
3、增加 mac 下的 shell 脚本 及 astyle.Darwin 程序。

+++ inc/simple/enum.h
@@ -59,7 +59,7 @@ private:
 	inline const char* const enum_to_string(const TYPE index)\
 {\
 	const static char* TYPE##EnumStrings[] = {VALUES ""};\
-	return (index >= 0 && index < sizeof(TYPE##EnumStrings)/sizeof(TYPE##EnumStrings[0]))?TYPE##EnumStrings[index]:"";\
+	return (index >= 0 && (unsigned int)(index) < sizeof(TYPE##EnumStrings)/sizeof(TYPE##EnumStrings[0]))?TYPE##EnumStrings[index]:"";\
 }
 #define	ENUM_NAME(Index)	(#Index),
 #define	ENUM_ITEM(VALUE)	(VALUE),
